+++ ImageAnalysis/image_analysis/offline_analyzers/__init__.py
@@ -6,6 +6,9 @@ experimental setups.
 """
 
 from image_analysis.offline_analyzers.beam_analyzer import BeamAnalyzer
+from image_analysis.offline_analyzers.frog_spectral_phase_analyzer import (
+    FrogSpectralPhaseAnalyzer,
+)
 from image_analysis.offline_analyzers.line_analyzer import LineAnalyzer
 from image_analysis.offline_analyzers.line_stitcher import LineStitcher
 from image_analysis.offline_analyzers.standard_1d_analyzer import Standard1DAnalyzer
@@ -13,6 +16,7 @@ from image_analysis.offline_analyzers.standard_analyzer import StandardAnalyzer
 
 __all__ = [
     "BeamAnalyzer",
+    "FrogSpectralPhaseAnalyzer",
     "LineAnalyzer",
     "LineStitcher",
     "Standard1DAnalyzer",